Composite refactor
