+++ Structural/Composite/src/interfaces/ProductComponent.ts
@@ -1,0 +1,3 @@
+export interface ProductComponent {
+  getPrice(): number;
+}
