+++ Structural/Composite/src/models/Product.ts
@@ -1,0 +1,9 @@
+import { ProductComponent } from "../interfaces/ProductComponent";
+
+export class Product implements ProductComponent {
+  constructor(private price: number) {}
+
+  getPrice() {
+    return this.price;
+  }
+}
